fix: render authored progress-outcome handles on unsupported workstations
Legacy onContinue/onRejection edges need hidden React Flow handles when
stopWords is absent so observer graphs render without endpoint errors.
