+++ docs/internal/processes/factory-workstation-relevant-files.md
@@ -85,6 +85,7 @@ work inputs.
 - Embedded observer-graph editor regressions should stay split across focused seams: keep handle ids plus shared label-fallback behavior in `ui/src/features/workflow-activity/lib/react-flow-current-activity-card-graph.test.ts`, keep draft-edge projection, outcome mapping, and removal filtering in `ui/src/features/workflow-activity/lib/react-flow-current-activity-card-draft-edges.test.ts`, keep draft-mutation and invalid-connection coverage in `ui/src/features/workflow-activity/hooks/react-flow-current-activity-card-editor-controllers.test.tsx`, and use `ui/src/features/workflow-activity/components/react-flow-current-activity-card.test.tsx` for the real DOM proof that shared-surface handle buttons and pending draft state stay visible without React Flow handle-attachment errors.
 - Cron/system-time customer-display regressions for the workflow-activity graph card should use `cronSystemTimeFactoryDefinition` / `dashboardSnapshotWithCronSystemTimeFactory` in `react-flow-current-activity-card.test.tsx` to prove rendered DOM omits `__system_time` work-type labels, expire utility workstations, and pending system-time states while customer cron workstations stay selectable.
 - Continue/reject graph handle visibility should key off `workstationSupportsProgressOutcomeRoutes` in `ui/src/features/current-factory-definition/lib/workstation-progress-outcome-routes.ts`: standard `MODEL_WORKSTATION` / `MODEL_INVOKE` processors need at least one trimmed `stopWords` entry, `REPEATER` workstations always qualify, and `CLASSIFIER_WORKSTATION` (or authored `classificationRoutes`) never qualify; factory-graph editor anchors, React Flow projection, and workflow-activity editor handles should all import this helper instead of re-deriving the rule.
+- When a standard processor lacks `stopWords` but factory JSON still has `onContinue`/`onRejection` edges, merge hidden progress-outcome handles via `mergeAuthoredProgressOutcomeConnectionAnchors` and `authoredProgressOutcomeSourceHandlesByWorkstationNodeId` so React Flow can render legacy routes without offering new continue/reject connections (`hidden: true`, `connectable: false` in editor connect mode).
 - Browser-visible continue/reject handle regressions for the factory graph editor belong in `ui/src/features/factory-graph-editor/components/factory-graph-editor-flow.stories.tsx` (`ProgressOutcomeRoutesWithoutStopWords`, `ProgressOutcomeRoutesWithStopWords`) with matching DOM coverage in `factory-graph-editor-flow.test.tsx` and Playwright assertions in `ui/scripts/verify-progress-outcome-route-handles-storybook-responsive.mjs` registered through `verify-import-export-storybook-responsive.mjs`.
 - Factory-graph editor connection anchors and connect validation live in `ui/src/features/factory-graph-editor/lib/factory-graph-editor-connections.ts`: pass `factoryGraphConnectionAnchorContext(workstation)` into `getFactoryGraphConnectionAnchors` / `getLocalizedFactoryGraphConnectionAnchors`, and pass `createFactoryGraphWorkstationResolver(factory.workstations)` into `buildFactoryGraphEdgeChangeFromConnection`, `connectFactoryGraphNodes`, and `projectFactoryGraphToReactFlow` (`workstationResolver`) so continue/reject anchors stay hidden and new connections are rejected for unsupported standard processors.
 - Workflow-activity graph editor handles should mirror the same capability rule in `ui/src/features/workflow-activity/lib/react-flow-current-activity-card-editor-handles.ts` via `resolveWorkstationConnectionAnchorContext(factory, factoryGraphNodeId)` passed into `buildSemanticGraphHandles`; put `progressOutcomeRouteWorkstation` on workstation node data for `react-flow-current-activity-card-viewport.tsx` connect validation, and resolve draft anchor clicks through `createFactoryGraphWorkstationResolver` in `react-flow-current-activity-card-editor-connections.ts`.
